Pull out instantiateAndConnectIbc
